backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
allure-pytest==2.13.5
allure-python-commons==2.13.5
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
frozenlist==1.5.0
greenlet==3.2.3
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.18.6
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
playwright==1.53.0
pluggy==1.5.0
propcache==0.2.1
pyee==13.0.0
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-check==2.5.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
python-rucaptcha==6.3.1
python-slugify==8.0.4
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tenacity==9.0.0
text-unidecode==1.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
